Extract test_helpers module
