+++ test_helpers.py
@@ -1,0 +1,37 @@
+"""Testing helper functions
+
+Copyright 2018 Spectre Team
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+import json
+from unittest.mock import MagicMock
+
+
+def returns(value) -> MagicMock:
+    return MagicMock(return_value=value)
+
+
+def many(values) -> MagicMock:
+    return MagicMock(side_effect=values)
+
+
+def throws(exception) -> MagicMock:
+    return MagicMock(side_effect=exception)
+
+
+def assert_valid_json(text: str):
+    try:
+        json.loads(text)
+    except json.JSONDecodeError as ex:
+        raise AssertionError(text) from ex
